requirements.txt
torch
torchvision
Pillow
numpy
tqdm
rasterio
matplotlib
scikit-image
scikit-learn
torchmetrics